Add frontend plus backend changes
This is a very large commit... as I did not commit (bad practice) during
the development of the second portion of this assignment.

Changes include the addition of a frontend for the API woo!

As well as fixing some of the API routes to adhere to RESTful standards,
adding request logging, and abstracting the database methods away from
the handlers themselves.